fix: 切り出しすぎたものを元に戻す対応
