+++ src/components/templates/CalcStatsTemplate.tsx
@@ -22,7 +22,6 @@ import { EffortValueField } from '@/components/organisms/EffortValueField'
 import { IndividualValueField } from '@/components/organisms/IndividualValueField'
 import { RealNumberField } from '@/components/organisms/RealNumberField'
 import { StatsTableHeader } from '@/components/organisms/StatsTableHeader'
-import { useStatsMutators } from '@/store/statsState'
 
 type Props = {
   buttonText: string
@@ -125,7 +124,22 @@ export const CalcStatsTemplate = (props: Props) => {
     updateStats(newStats)
   }
 
-  const { updateEffortValue } = useStatsMutators()
+  /**
+   * 努力値を更新する
+   *
+   * @param {number | ''} effortValue 努力値
+   * @param {number} statsIndex ステータス番号
+   * @return {void}
+   */
+  const updateEffortValue = (effortValue: number | '', statsIndex: number) => {
+    const newStats = stats.map((stat, index) => {
+      if (index === statsIndex) {
+        return { ...stat, effortValue }
+      }
+      return stat
+    })
+    updateStats(newStats)
+  }
 
   type EffortValue = {
     index: number
@@ -426,7 +440,11 @@ export const CalcStatsTemplate = (props: Props) => {
                 statsIndex={index}
                 updateIndividualValue={updateIndividualValue}
               />
-              <EffortValueField stats={stats} statsIndex={index} />
+              <EffortValueField
+                stats={stats}
+                statsIndex={index}
+                updateEffortValue={updateEffortValue}
+              />
               <RealNumberField
                 realNumber={realNumbers[index]}
                 stats={stats}
